Fix UI

+++ lib/src/view/book_management/library_diagrams/library_diagrams_page.dart
@@ -1,8 +1,11 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:htlib/src/model/admin_user.dart';
 import 'package:htlib/src/model/diagram_node.dart';
 import 'package:htlib/src/model/diagram_node_mode.dart';
+import 'package:htlib/src/services/admin_service.dart';
 import 'package:htlib/src/services/book/diagram_node_service.dart';
 import 'package:htlib/src/services/state_management/diagram_notifier/diagram_end_drawer_notifier.dart';
 import 'package:htlib/src/services/state_management/diagram_notifier/diagram_bottom_bar_notifier.dart';
@@ -21,18 +24,21 @@ class LibraryDiagram extends StatefulWidget {
 
 class _LibraryDiagramState extends State<LibraryDiagram> {
   DiagramNodeService _service = DiagramNodeService();
-  TransformationController _transformationController =
-      TransformationController();
+  TransformationController _transformationController = TransformationController();
 
   LibraryConfig get config => context.read();
   DiagramEndDrawerNotifier _nodeNotifier = DiagramEndDrawerNotifier();
   DiagramBottomBarNotifier _bottomBarNotifier = DiagramBottomBarNotifier();
 
   bool isInit = false;
-
+  AdminService? adminService;
   @override
   void initState() {
     super.initState();
+    try {
+      adminService = Get.find<AdminService>();
+    } catch (e) {}
+
     _service.init().then((_) {
       _nodeNotifier.changeNode(_service.anchor);
       _nodeNotifier.addListener(() {
@@ -66,10 +72,8 @@ class _LibraryDiagramState extends State<LibraryDiagram> {
       boundaryMargin: EdgeInsets.all(double.infinity),
       child: Padding(
         padding: EdgeInsets.symmetric(
-          horizontal:
-              max((MediaQuery.of(context).size.width - config.width) / 2, 0),
-          vertical:
-              max((MediaQuery.of(context).size.height - config.height) / 2, 0),
+          horizontal: max((MediaQuery.of(context).size.width - config.width) / 2, 0),
+          vertical: max((MediaQuery.of(context).size.height - config.height) / 2, 0),
         ),
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
@@ -99,8 +103,7 @@ class _LibraryDiagramState extends State<LibraryDiagram> {
                         _service.editNode(
                           node.copyWith(bookList: [...node.bookList, id]),
                         );
-                        _bottomBarNotifier
-                            .setUnsortedBookList(_service.unsortedBookList);
+                        _bottomBarNotifier.setUnsortedBookList(_service.unsortedBookList);
                         setState(() {});
                       },
                       onAddNewDirection: (direction) {
@@ -134,40 +137,48 @@ class _LibraryDiagramState extends State<LibraryDiagram> {
 
     return MultiProvider(
       providers: [
-        ChangeNotifierProvider<DiagramEndDrawerNotifier>.value(
-            value: _nodeNotifier),
-        ChangeNotifierProvider<DiagramBottomBarNotifier>.value(
-            value: _bottomBarNotifier),
+        ChangeNotifierProvider<DiagramEndDrawerNotifier>.value(value: _nodeNotifier),
+        ChangeNotifierProvider<DiagramBottomBarNotifier>.value(value: _bottomBarNotifier),
       ],
       child: Scaffold(
         appBar: AppBar(
           leading: IconButton(
             icon: Icon(Icons.arrow_back),
+            color: Theme.of(context).colorScheme.onPrimary,
             onPressed: () => Navigator.pop(context),
           ),
-          title: Text("Sơ đồ thư viện"),
+          title: Text(
+            "Sơ đồ thư viện",
+            style: Theme.of(context).appBarTheme.titleTextStyle!.copyWith(
+                  color: Theme.of(context).colorScheme.onPrimary,
+                ),
+          ),
+          backgroundColor: Theme.of(context).colorScheme.primary,
           actions: [
-            IconButton(
-              icon: Icon(Icons.delete),
-              tooltip: "Xóa nút mới nhất",
-              onPressed: () {
-                bool success = _service.removeNode();
-                if (success) {
-                  setState(() {});
-                } else {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(
-                      content: Text("Không thể xóa nút gốc"),
-                      behavior: SnackBarBehavior.floating,
-                    ),
-                  );
-                }
-              },
-            ),
+            if (adminService != null && adminService!.currentUser.value.adminType == AdminType.librarian)
+              IconButton(
+                icon: Icon(Icons.delete),
+                color: Theme.of(context).colorScheme.onPrimary,
+                tooltip: "Xóa nút mới nhất",
+                onPressed: () {
+                  bool success = _service.removeNode();
+                  if (success) {
+                    setState(() {});
+                  } else {
+                    ScaffoldMessenger.of(context).showSnackBar(
+                      SnackBar(
+                        content: Text("Không thể xóa nút gốc"),
+                        behavior: SnackBarBehavior.floating,
+                      ),
+                    );
+                  }
+                },
+              ),
             Builder(
               builder: (context) => Padding(
                 padding: EdgeInsets.only(right: Insets.sm),
                 child: IconButton(
+                  color: Theme.of(context).colorScheme.onPrimary,
                   icon: Icon(Icons.settings),
                   onPressed: () {
                     _nodeNotifier.settings();
@@ -185,10 +196,11 @@ class _LibraryDiagramState extends State<LibraryDiagram> {
             : Stack(
                 children: [
                   libraryMap(),
-                  Align(
-                    alignment: Alignment.bottomCenter,
-                    child: HtlibBottomBar(),
-                  ),
+                  if (adminService != null && adminService!.currentUser.value.adminType == AdminType.librarian)
+                    Align(
+                      alignment: Alignment.bottomCenter,
+                      child: HtlibBottomBar(),
+                    ),
                 ],
               ),
       ),
